Move cap_max_workers_in_pool to test/parallel_testsuite.py. NFC
This is the only remaining caller.
