+++ tools/shared.py
@@ -128,17 +128,6 @@ def returncode_to_str(code):
   return f'returned {code}'
 
 
-def cap_max_workers_in_pool(max_workers, is_browser):
-  if is_browser:
-    # TODO experiment with this number. In browser tests we'll be creating
-    # a chrome instance per worker which is expensive.
-    max_workers = int(max_workers / 2)
-  # Python has an issue that it can only use max 61 cores on Windows: https://github.com/python/cpython/issues/89240
-  if WINDOWS:
-    return min(max_workers, 61)
-  return max_workers
-
-
 def run_multiple_processes(commands,
                            env=None,
                            route_stdout_to_temp_files_suffix=None,
